split test data better
